requirements.txt
# 基础依赖

# 核心库
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
transformers>=4.30.0
huggingface-hub>=0.16.0

# 向量检索
faiss-cpu>=1.7.4

# 模型相关
ultralytics>=8.0.0
clip-anytorch>=2.5.2

# 工具库
numpy>=1.24.0
Pillow>=9.5.0
scipy>=1.10.0
matplotlib>=3.7.0

# 可选：Web UI
gradio>=3.40.0
streamlit>=1.20.0

# 可选：DeepDanbooru
deepdanbooru>=1.0.0

# 可选：Pixiv采集
pixivpy3>=3.7.0
requests>=2.31.0

flask
